requirements.txt
appdirs==1.4.4
Appium-Python-Client==1.0.2
certifi==2020.11.8
chardet==3.0.4
distlib==0.3.1
field==0.2.0
filelock==3.0.12
idna==2.10
importlib-metadata==3.1.0
mysqlclient==2.0.1
numpy==1.19.4
Pillow==8.0.1
Pygments==2.7.2
Pypubsub==4.0.3
pywin32==228
requests==2.25.0
robotframework==3.1.2
robotframework-pythonlibcore==2.1.0
robotframework-ride==1.7.4.2
robotframework-seleniumlibrary==4.5.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.11
virtualenv==20.2.1
wxPython==4.0.7.post2
zipp==3.4.0
